refactor(domain): 将场景标签数据类型从Byte改为Integer
- 修改Record实体类中sceneType字段类型为Integer
- 修改RecordAddDTO中sceneType字段类型为Integer
- 修改RecordListVO中sceneType字段类型为Integer

+++ src/main/java/com/miqroera/biosensor/domain/model/Record.java
@@ -1,10 +1,6 @@
 package com.miqroera.biosensor.domain.model;
 
-import com.baomidou.mybatisplus.annotation.FieldFill;
-import com.baomidou.mybatisplus.annotation.IdType;
-import com.baomidou.mybatisplus.annotation.TableField;
-import com.baomidou.mybatisplus.annotation.TableId;
-import com.baomidou.mybatisplus.annotation.TableName;
+import com.baomidou.mybatisplus.annotation.*;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
 import lombok.Setter;
@@ -67,7 +63,7 @@ public class Record implements Serializable {
      * 场景标签（0 未选 1 空腹 2 餐后 3 运动后）
      */
     @Schema(description = "场景标签（0 未选 1 空腹 2 餐后 3 运动后）")
-    private Byte sceneType;
+    private Integer sceneType;
 
     /**
      * 浓度值 (ppb)
